Update fib
Comment week 1 Fibonacci

My name is Lucia, so the first and last letter of my name (L + A = 12 + 1) give the number 13. The 13th Fibonacci number is 233

Comment week 2 
My surname is Saura
The first letter S is number 83
The last letter a is number 97
Fibonacci number 180 is 18547707689471986212190138521399707760
 
The (ord) returns an integer representing the Unicode code point of the character 'S' and of the character 'a'
 
Then both numbers are summed and the last part shows the Fibonacci number for the previous result (in my case 180).